Restore HCI_MON in bluetooth.py
